requirements.txt
openai
SpeechRecognition
pyaudio
websockets
python-dotenv
pydantic
pyttsx3
colorlog
